requirements.txt
discord-py-interactions>=5.11.0
mcstatus>=11.1.1
python-dotenv>=1.0.1
